cast

+++ game/src/Game.cpp
@@ -688,7 +688,7 @@ void Game::handleBotUpdates(const HandlerArgs &server) {
         auto &player = state.players[state.turn];
 
         BotDecisionState decision_state = {
-            .id = state.turn,
+            .id = (int)state.turn,
             .name  = (BotName)player.bot_enum,
             .phase = state.phase,
             .trump = state.trump,
